+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 -   use [.prettierignore](https://github.com/prettier/prettier/releases/tag/1.6.0) file instead of a complicated glob
 
+-   prettier: trailingComma='all' for Node.js 8+
+
 
 ## 3.0.0 - 2017-09-11